requirements.txt
absl-py==1.4.0
alembic==1.12.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clearml==1.12.2
click==8.1.7
cloudpickle==2.2.1
cmaes==0.10.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.1
cucim==23.8.0
cycler==0.11.0
databricks-cli==0.17.7
dcm2niix==1.0.20220715
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.3
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.11.0
fire==0.5.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.9.1
furl==2.1.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
imagecodecs==2023.9.18
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
miutil==0.12.0
mlflow==2.7.1
monai==1.2.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nni==3.0
notebook==7.0.3
notebook_shim==0.2.3
nptyping==2.5.0
numpy==1.26.0
nvidia-ml-py==12.535.108
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.15.1
openslide-python==1.1.2
optuna==3.3.0
orderedmultidict==1.0.1
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
prettytable==3.9.0
progressbar2==4.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydicom==2.4.3
Pygments==2.16.1
PyJWT==2.4.0
pynrrd==1.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==3.8.0
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.11
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
querystring-parser==1.2.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
schema==0.7.5
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sentry-sdk==1.31.0
setproctitle==1.3.2
setuptools==56.0.0
SimpleITK==2.3.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.18
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.21.3
typeguard==4.1.2
types-PyYAML==6.0.12.11
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wandb==0.15.10
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
wheel==0.41.2
widgetsnbextension==4.0.9
xnat==0.5.1
zipp==3.17.0
